+++ locksmith/src/operations/metadataOperations.ts
@@ -10,6 +10,8 @@ import * as lockOperations from './lockOperations'
 import { Attribute } from '../types'
 import metadata from '../config/metadata'
 import { getDefaultLockData } from '../utils/metadata'
+import { UserTokenMetadata } from '../models'
+import logger from '../logger'
 
 interface IsKeyOrLockOwnerOptions {
   userAddress?: string
@@ -246,6 +248,36 @@ export const getKeysMetadata = async ({
 
   const mergedData = await Promise.all(mergedDataList)
   return mergedData.filter(Boolean)
+}
+
+export const getUserProtectedMetadata = async ({
+  lockAddress,
+  userAddress,
+}: {
+  lockAddress: string
+  userAddress: string
+}) => {
+  const userTokenMetadataRecord = await UserTokenMetadata.findOne({
+    where: {
+      tokenAddress: lockAddress,
+      userAddress,
+    },
+  })
+
+  if (userTokenMetadataRecord) {
+    logger.info(
+      'Found the relevant token metadata',
+      userTokenMetadataRecord?.data
+    )
+
+    const protectedData = Normalizer.toLowerCaseKeys({
+      ...userTokenMetadataRecord?.data?.userMetadata?.protected,
+    })
+
+    return protectedData
+  }
+
+  return null
 }
 
 export const getLockMetadata = async ({
